requirements.txt
langchain-core>=0.2.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.5.0